store version history as utc; add self-heal logic
